+++ packages/core/src/createBox.ts
@@ -1,6 +1,0 @@
-import { system } from '@xstyled/system'
-
-export function createBox(): (string | typeof system)[] {
-  return [`&&{`, system, `}`]
-}
-createBox.meta = system.meta