+++ packages/styled-components/src/preflight.ts
@@ -1,6 +1,0 @@
-import { createGlobalStyle } from 'styled-components'
-import { getPreflightStyles } from '@xstyled/system'
-
-export const Preflight = createGlobalStyle(({ theme }) =>
-  getPreflightStyles(theme),
-)