+++ packages/xstyled-core/src/createBox.ts
@@ -1,0 +1,6 @@
+import { system } from '@agriffis/xstyled-system'
+
+export function createBox(): (string | typeof system)[] {
+  return [`&&{`, system, `}`]
+}
+createBox.meta = system.meta