+++ packages/xstyled-styled-components/src/preflight.ts
@@ -1,0 +1,6 @@
+import { createGlobalStyle } from 'styled-components'
+import { getPreflightStyles } from '@agriffis/xstyled-system'
+
+export const Preflight = createGlobalStyle(({ theme }) =>
+  getPreflightStyles(theme),
+)